Rename and test resource authorization on `NavsQuery`.
